% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{clusterLouvain}
\alias{clusterLouvain}
\title{Louvain clustering}
\usage{
clusterLouvain(relations)
}
\arguments{
\item{relations}{A relations matrix}
}
\value{
The community membership data frame
}
\description{
Louvain clustering
}
