% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dedup_links}
\alias{dedup_links}
\title{Remove duplicate links}
\usage{
dedup_links(links)
}
\arguments{
\item{links}{A numeric matrix of network edges}
}
\value{
The matrix of edges with duplicates removed
}
\description{
Remove duplicate links
}
