% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastPG.R
\name{fastCluster}
\alias{fastCluster}
\title{Fast approximate clustering}
\usage{
fastCluster(data, k = 30, num_threads = 1)
}
\arguments{
\item{data}{A numeric matrix of rows to cluster based on similarity of.
column features.}

\item{k}{Expected number of clusters.}

\item{num_threads}{Number of threads to use. Defaults to 1.}
}
\value{
Returns a list with two elements:
\itemize{
\item \code{modularity} - \link{TODO}
\item \code{communities} - An integer vector where the nth element is the nth
row in the input matrix. Its value is the cluster that row has been assigned to.
}
}
\description{
\link{TODO: find way to set default threads to max}
\link{TODO: how to estimat k? Sensitivity if wrong?}
\link{TODO: Description of algorithm; references}
}
\examples{
# [TODO]

}
