% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{fcsMatrix}
\alias{fcsMatrix}
\title{Load the data matrix from an FCS file}
\usage{
fcsMatrix(data, signalColumns = NULL, drop = TRUE, sample = FALSE,
  force = FALSE)
}
\arguments{
\item{data}{Can be a one element character vector giving the FCS file to
load or URL to download from, or a connection to
read the file from, or a flowCore::flowFrame object.}

\item{signalColumns}{The columns to keep.
Either a vector of names or integers (positions). By default will keep
all columns, which is probably not what you want. Any column not found
causes an error, unless force = TRUE is set, in which case missing
columns will be ignored with a warning.}

\item{drop}{Drop rows containing NA values.
By default is \code{TRUE}; An \code{NA} in any column will drop that row from the
matrix. Can be set to \code{FALSE} to keep all rows, or a vector of column
names or positions to check. Applied before any columns are dropped by
\code{signalColumns=}. Any column not found causes an error, unless force=
TRUE is set, in which case missing columns will be ignored with a
warning.}

\item{sample}{Select only this many rows of data.
Default is all data rows. If \code{sample} exceeds the number of rows of data,
will ignore with a warning. Applied after any rows are dropped due to
\code{NA} values.}

\item{force}{Try and process anyway, despite errors.}
}
\value{
A matrix of measurements for event rows (cells) x feature columns
(signal channels).
}
\description{
Load the data matrix from an FCS file
}
