% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexStringToRaw}
\alias{hexStringToRaw}
\title{Convert a hex string to a raw vector}
\usage{
hexStringToRaw(x, sep = NA_character_)
}
\arguments{
\item{x}{A hex string
Besides delimiters, may only contain case insensitive hex digits, i.e.
0-9, a-f, and A-F. Must have an even number of hex digits, and if
delimiters are present, must split the string into 2 digit pairs.

Converts \code{NULL}, \code{character(0)} and \code{""} to \code{NULL}, \code{raw(0)} and \code{raw(0)}
respectively.}

\item{sep}{A delimiter separating hex digit pairs.
By default is NA_character_, meaning no delimiter is present. The
delimiter is ignored if input is NULL, character(0), "", or a single two
digit value. Note that a string ends in delimiters, the last delimiter
is ignored, but it is an error for a string to begin with a delimiter.
This is an artifact of the way R implements strsplit().}
}
\value{
A raw vector equivalent to the hex string pairs.
}
\description{
Convert a hex string to a raw vector
}
\examples{
hexStringToRaw( "01ae03ff" )
#> 01 ae 03 ff

hexStringToRaw( "01 ae 03 FF", sep= " " )
#> 01 ae 03 ff

hexStringToRaw( "01::ae::03::", sep= "::" )
#> 01 ae 03

hexStringToRaw( NULL )
#> NULL

hexStringToRaw( "" )
#> raw(0)

hexStringToRaw( character(0) )
#> raw(0)

}
