% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{hnsw}
\alias{hnsw}
\title{Hierarchical navigable small world relations map}
\usage{
hnsw(dat, k = 30, nThread = 1)
}
\arguments{
\item{dat}{The data matrix to analyze}

\item{k}{The estimated number of clusters, defaults to 30}

\item{nThread}{The number of threads to use, defaults to 1}
}
\value{
The relations matrix
}
\description{
Hierarchical navigable small world relations map
}
