% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{loadFCS}
\alias{loadFCS}
\title{Read in an FCS file}
\usage{
loadFCS(fcsFile)
}
\arguments{
\item{fcsFile}{The FCS format file to read in}
}
\value{
The data matrix from an FCS file
}
\description{
Read in an FCS file
}
