% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{outputForGrapallo}
\alias{outputForGrapallo}
\title{Output data to file for running Grapallo}
\usage{
outputForGrapallo(relations, file, k,
  nThread = data.table::getDTthreads(FALSE))
}
\arguments{
\item{relations}{}

\item{file}{}

\item{k}{}

\item{nThread}{}
}
\description{
Output data to file for running Grapallo
}
