% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{outputForGrappolo}
\alias{outputForGrappolo}
\title{Output data to file for running Grappolo}
\usage{
outputForGrappolo(relations, outFile, k,
  nThread = data.table::getDTthreads(FALSE))
}
\arguments{
\item{relations}{The relations matrix to output}

\item{outFile}{The name of the file to output}

\item{k}{The number of clusters}

\item{nThread}{The number of cpus to use}
}
\value{
Returns the name of the file written
}
\description{
Output data to file for running Grappolo
}
