% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{parseCli}
\alias{parseCli}
\title{Parse command line}
\usage{
parseCli(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{A character vector of command line argument strings, usually
\code{commandArgs( trailingOnly= TRUE )}}
}
\value{
A list of options and their values, either set from the passed
in args= vector or by default.
}
\description{
Parse command line
}
