% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_jce}
\alias{rcpp_parallel_jce}
\title{Parallel Jaccard similarity index}
\usage{
rcpp_parallel_jce(mat)
}
\arguments{
\item{mat}{A numeric matrix of values}
}
\value{
A numeric matrix of values
}
\description{
Parallel Jaccard similarity index
}
