% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redownload}
\alias{redownload}
\title{(Re)download a file with verification}
\usage{
redownload(url, file = basename(url), sha256 = NA_character_)
}
\arguments{
\item{url}{The URL specifying the file to download.}

\item{file}{The name (path) to download the file to.
Defaults to saving into the current directory a file named the same as
the basename of the url. Using \code{file= tempfile()} lets the system clean
up the file at the end of the session.}

\item{sha256}{The sha256 checksum expected from the downloaded file, as a
character string with only hex digits. Case insensitive. Defaults to
na_character_, which means don't check do checksum validation.}
}
\value{
The filename (invisibly)

Note: regardless of whether a checksum succeeds, fails, or is skipped, the
file is left as downloaded.
}
\description{
Download a file from a URL and optionally verify the file matches a given
sha256 checksum. Will default to the url basename, but a different name can
be specified. Will only re-download a file that does not exist, but will
test a pre-existing file's checksum. If no checksum is provided when a file
pre-exists, a warning is given as the file present could be unrelated. It
is a fatal error if the checksum fails.
}
\examples{
\dontrun{
  checksum <- "ea8fac7c65fb589b0d53560f5251f74f9e9b243478dcb6b3ea79b5e36449c8d9"
  toFile <- tempfile()
  redownload("http://www.example.com", file= toFile, sha256= checksum)
}
}
