% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{runGrapallo}
\alias{runGrapallo}
\title{Run Grapallo parallel louvain clustering}
\usage{
runGrapallo(file, exec = "./bin/driverForGraphClustering", f = 8,
  c = 1, o = TRUE, v = TRUE)
}
\arguments{
\item{file}{}

\item{exec}{}

\item{f}{}

\item{c}{}

\item{o}{}

\item{v}{}
}
\description{
Run Grapallo parallel louvain clustering
}
