% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{runGrappolo}
\alias{runGrappolo}
\title{Run Grappolo parallel louvain clustering}
\usage{
runGrappolo(file, exec = "./bin/driverForGraphClustering", f = 8,
  c = 1, o = TRUE, v = TRUE)
}
\arguments{
\item{file}{The file to cluster}

\item{exec}{Grappolo executable, either on the path or the full path.}

\item{f}{The file type being read, as an integer. Defaults to 8 (SNAP format).
See \code{-f} in Grappolo}

\item{c}{An integer, defaults to 1. See \code{-c} in Grappolo}

\item{o}{Output? A boolean value, defaults to TRUE. See \code{-o} in Grappolo}

\item{v}{A boolean value, defaults to TRUE. See \code{-v} in Grappolo}
}
\value{
The exits status of the external command running Grappolo. Should
be 0 for success.
}
\description{
Run Grappolo parallel louvain clustering
}
