% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{hnsw}
\alias{hnsw}
\title{Hierarchical navigable small world relations map}
\usage{
hnsw(dat, k = 30, nThread = 5)
}
\arguments{
\item{dat}{}

\item{k}{}

\item{nThread}{}
}
\description{
Hierarchical navigable small world relations map
}
