% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallel_louvain}
\alias{parallel_louvain}
\title{Parallel Louvain clustering}
\usage{
parallel_louvain(links)
}
\arguments{
\item{links}{A numeric matrix of network edges}
}
\value{
A list with two elements
\itemize{
\item \code{modularity} -
\item \code{communities} - A vector where the i'th value is the
cluster number that the i'th node in the links matrix has been assigned to.
}
}
\description{
Parallel Louvain clustering
}
