% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPG.R
\name{runGrappolo}
\alias{runGrappolo}
\title{Run Grappolo parallel louvain clustering}
\usage{
runGrappolo(file, exec = "driverForGraphClustering", f = 8,
  c = FALSE, o = FALSE, v = FALSE, s = FALSE, r = FALSE,
  m = 100000L, d = 0.01, l = 1e-06)
}
\arguments{
\item{file}{The file to cluster}

\item{exec}{Grappolo executable, either on the path or the full path.}

\item{f}{The file type being read, as an integer. Defaults to 8 (SNAP format).
See \code{-f} in Grappolo}

\item{c}{An integer, defaults to FALSE. See \code{-c} in Grappolo}

\item{o}{A boolean value, defaults to FALSE. See \code{-o} in Grappolo}

\item{v}{A boolean value, defaults to FALSE. See \code{-v} in Grappolo}

\item{s}{A boolean value, defaults to FALSE. See \code{-s} in Grappolo}

\item{r}{A boolean value, defaults to FALSE. See \code{-r} in Grappolo}

\item{m}{An integer value, defaults to 100,000. See \code{-m} in Grappolo}

\item{d}{A numeric value, defaults to 0.01. See \code{-d} in Grappolo}

\item{l}{A numeric value, defaults to 0.000,001 See \code{-l} in Grappolo}
}
\value{
The exits status of the external command running Grappolo. Should
be 0 for success.
}
\description{
Run Grappolo parallel louvain clustering
}
